+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * adding `is/only-tabbable` to identify elements that cannot be focused by script, but by keyboard
 * removing "dimension check" from `is/visible` - [issue #14](https://github.com/medialize/ally.js/issues/14)
 * `query/focusable` no longer considers `<html>` and `<body>` focusable - [issue #31](https://github.com/medialize/ally.js/issues/31)
+* adding `element/disabled` (refactored from `focus/disabled`) - [issue #33](https://github.com/medialize/ally.js/issues/33)
 
 ### Breaking Changes